Break up QuerySummaryPanel
