+++ packages/query-composer/src/components/FieldButton/index.ts
@@ -21,4 +21,4 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-export {FieldButton} from './FieldButton';
+export {FieldButton, BackPart, CloseIconStyled} from './FieldButton';
