+++ packages/query-composer/src/components/QuerySummaryPanel/FieldListDiv.tsx
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ *  LICENSE file in the root directory of this source tree.
+ */
+
+import styled from 'styled-components';
+
+export const FieldListDiv = styled.div`
+  display: flex;
+  gap: 2px;
+  flex-direction: column;
+`;
